manager: fix disconnecting testers
When the controller sends the bug list, the controller disconnects
straight after. As a result, the code in manager for processing the
bugs gets cancelled in weird way which results in successful sending
to tester, but somehow kills the connection.

By moving the bugs list processing to separate coroutine which is
ran in background, the code doesn't fail when controller disconnects
and it looks like it works now.

Hopefully this works now. In case it isn't, you can view this
occurrence by using `./controller.py -b` and seeing in manager's log
"sent to ... bugs ..." and then "normal connection closed" for the
tester connection.

Signed-off-by: Arthur Zamarin <arthurzam@gentoo.org>